# [Exec](expr) Opt the compound pred performace